Add files via upload
my_app_module.py is a Python file custom module that has a custom function created and imported to the notebook file. Raw_Rewiews.csv is the main dataset. Output1 is the prediction of labels of negative reviews (21-50) age group and only for the tops and dresses product type.
output 2 is the prediction of labels of positive reviews of all (but 30 percent random samples).